// src/components/api/kiti/schema.ts
export interface paths {
// ...
  '/api/events/': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    get: operations['api_events_list']
    put?: never
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    post: operations['api_events_create']
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  '/api/events/{id}/': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    get: operations['api_events_retrieve']
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    put: operations['api_events_update']
    post?: never
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    delete: operations['api_events_destroy']
    options?: never
    head?: never
    /** @description API endpoint for events.
     *
     *     list:
     *     Returns a list of all the existing events.
     *
     *     retrieve:
     *     Returns the given event.
     *
     *     create:
     *     Creates a new event instance.
     *
     *     update:
     *     Updates a given event.
     *
     *     partial_update:
     *     Updates a given event.
     *
     *     destroy:
     *     Removes the given event. */
    patch: operations['api_events_partial_update']
    trace?: never
  }
// ...
}
export type webhooks = Record<string, never>
export interface components {
  schemas: {
// ...
    Event: {
      readonly id: number
      name: string
      description?: string
      /**
       * Start date
       * Format: date
       */
      date_start: string
      /**
       * End date
       * Format: date
       */
      date_end: string
      location: string
      organization: number
      readonly organization_info: components['schemas']['Organization']
      public?: boolean
      readonly competitions: components['schemas']['CompetitionInfo'][]
      /** Edit lock */
      locked?: boolean
      approved?: boolean
      readonly permissions: string
      /** Competition categories */
      categories?: string
      optional_dates?: string
      /** Format: uri */
      web_page?: string
      /**
       * Invitation URL
       * Format: uri
       */
      invitation?: string
      /** Generic notes */
      notes?: string
      /** International competition */
      international?: boolean
      /** Safety plan exists */
      safety_plan?: boolean
      /** Terms and Conditions agreement */
      toc_agreement?: boolean
    }
// ...
    PaginatedEventList: {
      /** @example 123 */
      count: number
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?page=4
       */
      next?: string | null
      /**
       * Format: uri
       * @example http://api.example.org/accounts/?page=2
       */
      page: number
      limit: number
      previous?: string | null
      results: components['schemas']['Event'][]
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}
export type $defs = Record<string, never>
export interface operations {
  api_areas_list: {
    parameters: {
      query?: {
        /** @description Number of results to return per page. */
        limit?: number
        /** @description The initial index from which to return the results. */
        offset?: number
      }
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': components['schemas']['PaginatedAreaList']
        }
      }
    }
  }
// ...
}